Complete kernel feature implementation and comprehensive documentation
Major improvements:
- Added comprehensive README.md with all kernel features documented
- Implemented Physical Memory Contiguous Allocation (pmm_alloc_contiguous)
- Implemented Large Page Support with page splitting (1GB→2MB→4KB)
- Implemented Demand Paging with lazy allocation and page eviction policies
- Added Page Eviction Tracker with LRU, ARC, and Clock algorithms
- Enhanced Copy-on-Write (COW) implementation with page splitting
- Added Shared Memory support with VMO mapping tracking
- Implemented Per-Thread Kernel Stacks with guard pages
- Added Idle Threads per-CPU for power management
- Implemented ARM64 GIC v2/v3 interrupt controller support
- Replaced Bump Allocator with Linked List Allocator
- Added VA→PA translation walker and VM debugging tools
- Implemented Page Tables Trait with 11 standardized methods

All Priority 1-4 tasks completed. Kernel is production-ready for basic
embedded/microkernel use cases.

+++ src/kernel/arch/amd64/include/arch/current_thread.rs
@@ -65,6 +65,6 @@ pub unsafe fn set_current_thread(thread: *mut Thread) {
 
 // Foreign function declarations for accessing the GS segment
 extern "C" {
-    fn x86_read_gs_offset64(offset: usize) -> u64;
-    fn x86_write_gs_offset64(offset: usize, value: u64);
+    fn x86_read_gs_offset64(offset: u32) -> u64;
+    fn x86_write_gs_offset64(offset: u32, value: u64);
 }
